requirements.txt
ibmcloudant~=0.0.43
python-dotenv~=0.19.2
requests~=2.27.1